requirements.txt
flet==0.80.5
firebase-admin==6.9.0
google-cloud-firestore==2.21.0
python-dotenv==1.1.1
python-dateutil==2.9.0.post0
requests==2.32.5
pytest==8.4.1
